requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
openai==1.3.5
pillow==10.1.0
python-magic==0.4.27